# dsd parsing: remove allocations during int and float parsing